in relation to commit 503326bf8735d8eab48d4ff0ab9c000ffa031dec reviewed handling of non canonical paths
instead of trigger execption when one or more / follows in a path, they are
now ignored and replaced by a single one, this address the case of trailing slash
that the commit referred in title broke (was not more accepted as valid path).

Only URL-like formated string, which have double slash at the begining, are the double slashed that are preserved.