fixing bug due to race condition between two threads leading dar/libdar in a dead-lock condition (but very unfrequently)
while restoring EA, in sequential read mode with parallel block compression, the underlying parallel_tronconneuse
(multi-threaded deciphering engine) could be access concurrently from the parallel block engine (N worker
threads plus the feeder thread (zip_below_read class) which once the compressed data has been read had
still a HEOF block flag to read before ending, and the main thread that returned from the EA restoration
step and skips forward to the next mark which also means reading on the underlying paralle_tronconneuse.

As the paralle_tronconneuse class is not expected to be read() by two different threads at the same time
depending on the order of operation this leads to this thread waiting for data from its worker while its
worker threads already reported EOF and are suspended on a barrier waiting form the parallel_tronconneuse
to rejoin them in order for all threads (read_below, the workers and parallel_tronconneuse) to be released
in sync.

The fix is to set the compression engine to suspend after having loaded EA in memory and make the
compression suspension in read-mode to flush the data (which means also ending the worker and compression
feeder thread).