fixing bug in sar::truncate()
bug occurs when slicing is used (espetially with tiny slice size) and one of the last
file under backup, was so badly compressed that resaving it uncompressed uses several slices
less than when stored compressed and not enough data to be wrote to after that file in the
archive (other file, catalogue, ending elastic-buffer, terminators) lead to recreate the
slice size reached before resaving this file uncompressed.

resaving uncompressed, led to properly remove the extra slices but the sar class was not resetting
the max slice seen and when dropping the terminator at "end of file" the sar object skipped at end
of file according to the non existing slice seen before truncated, which recreated that slice
to drop the terminator structure. This led a hole in slice numbering and unability
at reading time for dar to find the missing slice in that hole.