fixing bug in parallel_tronconneuse
if data was pending in from the worker and lus_data/lus_flag was not containing
the requested offset, the method parallel_tronconneuse::purge_unack_stop_order
flushed lus_data/lus_flags as expected, loaded the from the pending block from
workers and instead of continuing the lookup, was disabling it as if an error
was met in lus_data/lus_flags... the data in transit was this destroyed and needed
to be reload from the underlying object. However when this one is a sar on a
tuyau_global that cannot skip back, this operation fails and the data at the
requested offset was unaccessible, leading in some very specific context the
data comparison to fail for some files (need a external catalog and sequential
reading mode without compression set for the archive).